Remove dead line for "export PATH", move _SUB_ROOT"
The commented out line for "export PATH" was duplicated as a live line of code a few lines earlier. There was no apparent reason to keep the extra commented out copy. 

It's not clear why the "export _SUB_ROOT" line was commented out, but it gave the appearance that commenting it in would make it work. However, it doesn't work because $libexec_path is not yet defined. By moving this line slightly further down, it will work if it is commented in.